# docs/swagger.yaml
definitions:
...
  models.Device:
    properties:
      active:
        type: integer
      baseband_version:
        type: string
      build_number:
        type: string
      carrier:
        type: string
      cloud:
        type: integer
      config:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      hub:
        type: string
      id:
        type: integer
      imei:
        type: string
      model:
        type: string
      model_image_url:
        type: string
      number:
        type: string
      os_version:
        type: string
      platform:
        type: string
      port:
        type: integer
      priority:
        type: string
      serial:
        type: string
      server:
        type: string
      smart_call_hiya:
        type: integer
      sp_software_version:
        type: string
      ui_version:
        type: number
      updated_at:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        example: error
        type: string
    type: object
...
  models.Server:
    properties:
      cloud_device_type:
        type: string
      cloud_name:
        type: string
      cloud_state:
        type: string
      cloud_status:
        type: string
      cloud_type:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      ip:
        type: string
      server_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.UpdateServerRequest:
    properties:
      cloud_name:
        type: string
      cloud_state:
        type: string
      cloud_status:
        type: string
      cloud_type:
        type: string
      ip:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_login:
        type: string
      password:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
      verified_at:
        type: string
    type: object
  service.CreateUserRequest:
    properties:
      email:
        type: string
      is_active:
        type: boolean
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: API для работы с DC Analytics Service.
  termsOfService: http://swagger.io/terms/
  title: DC Analytics Service API
  version: "1.0"
paths:
...
  /v1/analytics/servers:
    get:
      description: Возвращает список всех серверов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Server'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: GetServers
      tags:
      - servers
  /v1/analytics/servers/{id}:
    get:
      description: Возвращает сервер по заданному ID
      parameters:
      - description: ID сервера
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Server'
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: GetServerByID
      tags:
      - servers
    put:
      consumes:
      - application/json
      description: Обновляет запись сервера с заданным ID. Передается JSON с полями
        для обновления.
      parameters:
      - description: ID сервера
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления сервера
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/models.UpdateServerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Server'
        "400":
          description: Неверный формат запроса или ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Сервер не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: UpdateServer
      tags:
      - servers
  /v1/analytics/servers/{id}/devices:
    get:
      consumes:
      - application/json
      description: Возвращает список устройств для сервера с указанным ID.
      parameters:
      - description: ID сервера
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Device'
            type: array
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: GetDevicesByServerID
      tags:
      - servers
...
  /v1/analytics/users:
    get:
      consumes:
      - application/json
      description: Возвращает массив пользователей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: GetUsers
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Принимает данные и создает пользователя в системе
      parameters:
      - description: Данные пользователя
        in: body
        name: userData
        required: true
        schema:
          $ref: '#/definitions/service.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Неверные данные для создания пользователя
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: CreateUser
      tags:
      - users
  /v1/analytics/users/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет пользователя по его ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь удален
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: DeleteUser
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Возвращает данные пользователя, если он существует
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: GetUserByID
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Введите JWT-токен в формате: "Bearer <ваш_токен>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"